增加回复评论

+++ src/main/java/com/kevin/community/controller/QuestionController.java
@@ -13,7 +13,7 @@ public class QuestionController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/question/{id}")
-    public String question(@PathVariable("id") Integer id,
+    public String question(@PathVariable("id") Long id,
                            Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         //增加阅读数
